requirements.txt
aiosqlite==0.22.1
annotated-doc==0.0.4
annotated-types==0.7.0
anthropic==0.88.0
anyio==4.13.0
APScheduler==3.11.2
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.7
click==8.3.1
cryptography==46.0.6
distro==1.9.0
docstring_parser==0.17.0
fastapi==0.135.3
google-api-core==2.30.2
google-api-python-client==2.193.0
google-auth==2.49.1
google-auth-httplib2==0.3.1
google-auth-oauthlib==1.3.1
googleapis-common-protos==1.74.0
h11==0.16.0
httpcore==1.0.9
httplib2==0.31.2
httpx==0.28.1
idna==3.11
iniconfig==2.3.0
jiter==0.13.0
jsonpatch==1.33
jsonpointer==3.1.1
langchain-anthropic==1.4.0
langchain-core==1.2.25
langgraph==1.1.5
langgraph-checkpoint==4.0.1
langgraph-checkpoint-sqlite==3.0.3
langgraph-prebuilt==1.0.9
langgraph-sdk==0.3.12
langsmith==0.7.25
oauthlib==3.3.1
orjson==3.11.8
ormsgpack==1.12.2
packaging==26.0
pluggy==1.6.0
proto-plus==1.27.2
protobuf==7.34.1
pyasn1==0.6.3
pyasn1_modules==0.4.2
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.20.0
pyparsing==3.3.2
pytest==9.0.2
python-dotenv==1.2.2
python-telegram-bot==22.7
PyYAML==6.0.3
requests==2.33.1
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
sniffio==1.3.1
sqlite-vec==0.1.9
starlette==1.0.0
tenacity==9.1.4
typing-inspection==0.4.2
typing_extensions==4.15.0
tzlocal==5.3.1
uritemplate==4.2.0
urllib3==2.6.3
uuid_utils==0.14.1
uvicorn==0.42.0
xxhash==3.6.0
zstandard==0.25.0
